+++ libs/plugins/modifiercompiler.default.php
@@ -18,7 +18,7 @@
  *
  * @return string with compiled code
  */
-function smarty_modifiercompiler_default($params)
+function smarty_modifiercompiler_default($params, Smarty_Internal_TemplateCompilerBase $compiler)
 {
     $output = $params[ 0 ];
     if (!isset($params[ 1 ])) {
@@ -26,7 +26,13 @@ function smarty_modifiercompiler_default($params)
     }
     array_shift($params);
     foreach ($params as $param) {
-        $output = '(($tmp = @' . $output . ')===null||$tmp===\'\' ? ' . $param . ' : $tmp)';
+
+        if ($compiler->syntaxMatchesVariable($output)) {
+            $output = '(empty(' . $output . ') ? ' . $param . ' : ' . $output . ')';
+        } else {
+            $output = '(($tmp = ' . $output . ')===null||$tmp===\'\' ? ' . $param . ' : $tmp)';
+        }
+
     }
     return $output;
 }